pyproject.toml
[tool.poetry]
name = "ccflow"
version = "0.1.1"
description = ""
authors = ["Roger Hurwitz <rogerhurwitz@inspirsmith.com>"]
readme = "README.md"
packages = [{include = "ccflow", from = "src"}]

[tool.poetry.dependencies]
python = "^3.11"
requests = "^2.31.0"


[tool.poetry.group.dev.dependencies]
pytest = "^7.4.0"
python-dotenv = "^1.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
